landing page
I'm thrilled to share my latest project as part of my internship—developing a landing page for Apna Courses for my Level 1-Task 1 in @codsoft. This experience has been a fantastic opportunity to merge creativity with essential web development skills, resulting in a sleek and user-friendly design.

✨ What I Learned:
🔹 Efficiently creating columns and section divisions
🔹 Strategically arranging items for an optimal layout
🔹 Crafting and styling headers and footers
🔹 Perfecting alignments, padding, and color schemes
🔹 Enhancing design with precise element management
🔹 Implementing CSS practices to prevent overlapping elements

This project has been a rewarding learning journey, and I'm excited to showcase the results. Check it out and share your feedback! 💻✨
#WebDevelopment #HTML #CSS #LandingPage #CodSoft #WebDesign #LearningByDoing #TechJourney